move f1 metric to metrics module
