+++ src/metrics/__init__.py
@@ -1,0 +1,1 @@
+from .f1 import F1Metric
